Moved existing tests to test folder.
