--- src/tests/demo-test.test.js
@@ -1,5 +1,0 @@
-describe('Example test', () => {
-  it('should pass test', () => {
-    expect(true);
-  });
-});
